- Pull out LatencyDetails and LatencyIndicator to separate files
